Add data loading for the language model
**Description**

This commit adds a sample data loading for the language model.

Consider the text of context length/block_ size 8: Before w
Encoded x: [14, 43, 44, 53, 56, 43, 1, 61]
        y: [43, 44, 53, 56, 43, 1, 61, 43]
For a GPT Model, we want context with every other word which comes "before" it. That is,

For 'B', e is the output --> x: [14] y:[43]
For 'Be', f is the output --> x: [14, 43] y:[44]
For 'Bef', o is the output and so on --> x: [14, 43, 44] y:[53]
Hence why targets are shifted by +1

But for Bi-Gram Model, we don't consider the context, instead we just predict the next character

B --> e
e --> f
f --> o